requirements.txt
django~=5.0.7
djangorestframework~=3.15.2
django-filter~=24.2
